fix(android): revert package name to com.anonymous.mobileclient to preserve EAS credentials
