+++ apps/mobile-client/android/.gitignore
@@ -14,5 +14,3 @@ local.properties
 
 # Bundle artifacts
 *.jsbundle
-local.properties
-.kotlin/
