--- apps/mobile-client/android/app/src/main/java/com/anonymous/mobileclient/WabridgePackage.kt
@@ -1,14 +1,0 @@
-package com.anonymous.mobileclient
-
-import com.facebook.react.ReactPackage
-import com.facebook.react.bridge.NativeModule
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.uimanager.ViewManager
-
-class WabridgePackage : ReactPackage {
-    override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> =
-        listOf(WabridgeModule(reactContext))
-
-    override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> =
-        emptyList()
-}
