+++ apps/mobile-client/android/app/src/main/res/values/colors.xml
@@ -1,5 +1,5 @@
 <resources>
-  <color name="splashscreen_background">#ffffff</color>
-  <color name="iconBackground">#E6F4FE</color>
+  <color name="splashscreen_background">#E5FAEF</color>
+  <color name="iconBackground">#E5FAEF</color>
   <color name="colorPrimary">#023c69</color>
 </resources>
