+++ apps/mobile-client/android/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-  <string name="app_name">mobile-client</string>
+  <string name="app_name">WACI</string>
   <string name="expo_splash_screen_resize_mode" translatable="false">contain</string>
 </resources>
